feat(emacs): Add a right-click menu to the Emacs mode
The right-click menu is computed from a list of functions that produce
menu items based on Lean server information. Each menu item is a
string and a closure to be invoked if the menu item is selected.

There are currently two menu features: "Find definition" works where
M-. does, and on holes, the menu includes the hole commands from the
server.