Modified overlays for the message boxes
The `C-c C-b` message boxes had a couple of quirks that this commit
addresses.  The first is that they were placed *after* comments that
followed a command.  For example, it might show up as

> #eval 1+2
> -- This was an example of eval
> │ 3

rather than

> #eval 1+2
> │ 3
> -- This was an example of eval

It seems the basic reason this happens is that `lean` itself gives a
very conservative end position for a message, after having consumed
both whitespace and comments.  The `lean-message-boxes--make-overlay`
already took into account whitespace, so this commit introduces a new
function `lean-message-boxes--in-comment` to give it the ability to
also take into account characters that are part of a comment, which
seems to be a widely used method that is reliable so long as
`font-lock-mode` is on.

The second quirk is that a command at the very end of a file would not
show its message box unless there was an additional newline.  This
seems to have been because `lean-message-boxes--make-overlay` would
try to add the overlay to beyond the end of the file.  Rather than
moving the point forward, this commit changes the behavior to prefix
the overlay's string with a newline, which has seemed to be more
reliable.  There are some other small changes to allow the cursor to
be in a sensible position with respect to the overlay.