Add option to turn of automatic syncing
Sometimes it is desirable to not have the server be synced on each
change. One example is when you are writing a complicated type signature
for a lemma. I have often found that I don't need any feedback until
I've finished writing it out because the type signature will be nonsense
until then.